+++ src/main/java/in/hocg/web/modules/security/SecurityConfig.java
@@ -1,15 +1,10 @@
 package in.hocg.web.modules.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.security.SecurityProperties;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -20,10 +15,10 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
  * Created by hocgin on 2017/10/25.
  * email: hocgin@gmail.com
  */
-@Configuration
-@EnableWebSecurity
-@Order(SecurityProperties.ACCESS_OVERRIDE_ORDER)
-@EnableGlobalMethodSecurity(prePostEnabled = true)
+//@Configuration
+//@EnableWebSecurity
+//@Order(2)
+//@EnableGlobalMethodSecurity(prePostEnabled = true)
 public class SecurityConfig extends WebSecurityConfigurerAdapter {
     
     @Autowired
@@ -56,7 +51,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers("/auth/**").permitAll()
                 // 允许匿名访问 后台登陆
                 .antMatchers("/admin/login",
-                        SecurityConstants.SIGN_UP_URL).permitAll()
+                        SecurityConstants.ADMIN_SIGN_UP_PAGE).permitAll()
                 
                 // 允许对于网站静态资源的无授权访问
                 .antMatchers(
@@ -74,15 +69,12 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                         "/**/*.woff2"
                 ).permitAll()
                 
-                .and().exceptionHandling().authenticationEntryPoint(new ILoginUrlAuthenticationEntryPoint(SecurityConstants.SIGN_UP_URL)).and()
-                
-                .authorizeRequests()
                 // 除上面外的所有请求全部需要鉴权认证
                 .anyRequest().authenticated().and()
         
                 .formLogin()
-                    .loginPage(SecurityConstants.SIGN_UP_URL)
-                    .failureUrl(String.format("%s?error=true", SecurityConstants.SIGN_UP_URL))
+                    .loginPage(SecurityConstants.ADMIN_SIGN_UP_PAGE)
+                    .failureUrl(String.format("%s?error=true", SecurityConstants.ADMIN_SIGN_UP_PAGE))
                     .permitAll().and()
                 .logout()
                     .logoutUrl("/admin/logout");