+++ src/main/java/in/hocg/web/modules/service/RoleService.java
@@ -26,4 +26,5 @@ public interface RoleService {
     
     void save(String id, String[] permissionIds, CheckError checkError);
     
+    void deleteAllByDepartmentIn(String... DepartmentId);
 }